파일 요구사항대로 수정 2
